+++ src/collective/eeafaceted/dashboard/tests/test_dashboardcollection.py
@@ -2,7 +2,9 @@
 """Setup/installation tests for this package."""
 
 from plone import api
+from collective.eeafaceted.collectionwidget.utils import getCollectionLinkCriterion
 from collective.eeafaceted.dashboard.testing import IntegrationTestCase
+from collective.eeafaceted.dashboard.utils import enableFacetedDashboardFor
 from zope.component import queryUtility
 from zope.schema.interfaces import IVocabularyFactory
 
@@ -17,13 +19,13 @@ class TestDashboardCollection(IntegrationTestCase):
             id='c1',
             type='Collection',
             title='Collection 1',
-            container=self.portal
+            container=self.portal.folder
         )
         self.dashboardcollection = api.content.create(
             id='dc1',
             type='DashboardCollection',
             title='Dashboard collection 1',
-            container=self.portal
+            container=self.portal.folder
         )
 
     def test_MetaDataFieldsVocabulary(self):
@@ -35,3 +37,9 @@ class TestDashboardCollection(IntegrationTestCase):
         self.assertFalse('select_row' in factory(self.collection).by_token.keys())
         # DashboardCollection
         self.assertTrue('select_row' in factory(self.dashboardcollection).by_token.keys())
+
+    def test_enableFacetedDashboardFor_with_default_UID(self):
+        """ """
+        collection_uid = self.collection.UID()
+        enableFacetedDashboardFor(self.portal.folder, default_UID=collection_uid)
+        self.assertEqual(getCollectionLinkCriterion(self.portal.folder).default, collection_uid)